Fix AOCC compiler build issues
- Lowers minimum clang version from 11.0 to 8.0 to comply with AOCC 2.0
- Adds a newline at the end of each file
- Fix implicit type conversion error by typecasting explicitly

+++ test/aoclfftz_corebench_utils.h
@@ -883,7 +883,7 @@ VOID prepare_input_data_f(VOID *input, INTP n, INTP stride, INT32 input_type)
         INTP shift = rand() % length;
         // scale the amplitude of the wave by `scale` times, scale range:
         // [0.0 5.0)
-        FLOAT scale = ((FLOAT)rand() / RAND_MAX) * 5.0;
+        FLOAT scale = ((FLOAT)rand() / (FLOAT)RAND_MAX) * 5.0;
         for (INTP i = 0; i < length; i++)
         {
             input_f[((i + shift) % length) * T_DATA_STRIDE] =
